Finished Items and Weapons
Weapon Class implements Item
String name         // the name of the Weapon
String description // the description of the Weapon
int damage          // the damage of the Weapon
boolean pickUp   // true if the Weapon can be picked up, false if not

public Weapons() - made a series of telescoping constructors

void setName() - set the name of a Weapon
String getName() - returns the name of a Weapon

void setDescription() - set the description of a Weapon
String getDescription() - returns the description of a Weapon

void useItem() - prints out "You ATTACK with" + name of the Weapon.
                - also prints out "You did " + damage of the Weapon.
                - this does not check if the target of the weapon is valid

String toString() - prints out "A " + name + ". " + description + " . It does " + damage + " damage."
                           - example " A Sword. A sharp blade. It does 10 damage."